Changing unity_ProbesOcclusion from float4 to half4.

+++ com.unity.render-pipelines.universal/ShaderLibrary/UnityInput.hlsl
@@ -111,7 +111,7 @@ real4 unity_WorldTransformParams; // w is usually 1.0, or -1.0 for odd-negative 
 half4 unity_LightData;
 half4 unity_LightIndices[2];
 
-float4 unity_ProbesOcclusion;
+half4 unity_ProbesOcclusion;
 
 // Reflection Probe 0 block feature
 // HDR environment map decode instructions
